revises the style of the groups Search and Todo
The Search group conflicted with the cursor colors and you just could not see
which was the match the cursor was on. We are going to switch to the slate blue
here.

Because of that we revise the background of the less common Todo notes. Since
they need to be apparent this patch introduces a strong bold black on yellow.