let Types stand out
The initial style for Type was influenced by Ruby. But editing C code wih no
highlighting for types does not feel right to me. Let's do it the other way
around, silence rubyConstant which are clear because they start with an upper
case letter and there are so many that they would become noisy, and highlight
types by default.

Let's see how it works for other languages.