revises the color of the cursor line
The color for the cursor line was meant to be just a tad less
darker than the original #00000 background, but the contrast
was no longer visible with the current softened black.

I have been looking for a dark cursor line that is visible
but blends well with the background, I like it to be just a
nuance, a subtle visual hint. To be honest I am still not
totally convinced by this shade of gray, I took it from

    https://www.w3schools.com/colors/colors_shades.asp

I believe I would prefer it to be a little darkish but by
now let's push this so that at least it will work and will
have aesthetics close enough to my taste.

Thanks to @eparreno for feedback on this.